refactor(companhias-aereas): movendo entidades para pacote model
Refs: #28
