+++ sistema-companhia-aerea/src/main/java/br/edu/ifto/sistemacompanhiaaerea/model/Estado.java
@@ -1,4 +1,4 @@
-package br.edu.ifto.sistemacompanhiaaerea;
+package br.edu.ifto.sistemacompanhiaaerea.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
